server/requirements.txt
quart==0.19.4
python-dotenv==1.0.1
quart_jwt_extended==0.1.0
pymongo==4.6.2
cerberus==1.3.5
bcrypt==4.1.2
langchain==0.1.16
sentence_transformers > /dev/null
langchain_experimental==0.0.57
SentencePiece==0.2.0
protobuf==5.26.1
pypdf==4.2.0
langchain-community==0.0.33
rank_bm25 > /dev/null
faiss-gpu==1.7.2
pytesseract==0.3.10
poppler-utils==0.1.0
opencv-python==4.9.0.80
unstructured[docx]==0.11.8
